==> movie_lens/01_preparation.ipynb <==
import datetime
import re

import IPython
import matplotlib as mpl
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import plotly as py
from plotly.offline import download_plotlyjs
import pylab
import scipy
import seaborn as sns
import sklearn
from sklearn import *
import statsmodels as sm

np.random.seed(1337)

%matplotlib inline

IPython.core.display.display(IPython.core.display.HTML(
    "<style>.container { width:90% !important; }</style>"))

sns.set(font_scale=1.3)
sns.set_style('whitegrid')
sns.set_palette(sns.color_palette('muted'))

pd.options.display.max_colwidth = 30
pd.options.display.float_format = '{:,.3f}'.format

py.offline.init_notebook_mode(connected=True)

## Movies

movies = pd.read_csv('data/ml-20m/movies.csv')
movies.info()

movies.columns = ['movie_id', 'title', 'genres']

year_pattern = re.compile(r'.*\((\d+)\)')
def parse_year(title):
    try:
        return int(year_pattern.match(title).group(1))
    except:
        return None
movies['year'] = movies.title.apply(parse_year)

all_genres = set()
def parse_genres(genres):
    ret = []
    for genre in genres.split('|'):
        genre = re.sub('[^a-z]', '', genre.lower())
        ret.append(genre)
        all_genres.add(genre)
    return ret
movies['parsed_genres'] = movies.genres.apply(parse_genres)
del movies['genres']
for genre in all_genres:
    movies['genre_{0}'.format(genre)] = movies.parsed_genres.apply(lambda genres: int(genre in genres))

movies.sample(10).T

## Tags

tags = pd.read_csv('data/ml-20m/tags.csv')
tags.info()

tags.columns = ['user_id', 'movie_id', 'tag', 'timestamp']

tags['year'] = tags.timestamp.apply(lambda ts: datetime.datetime.fromtimestamp(ts).year)

tags.sample(10)

## Ratings

ratings = pd.read_csv('data/ml-20m/ratings.csv')
ratings.info()

ratings.columns = ['user_id', 'movie_id', 'rating', 'timestamp']

ratings['year'] = ratings.timestamp.apply(lambda ts: datetime.datetime.fromtimestamp(ts).year)

ratings.sample(10)

## Write to disk

movies.to_csv('data/movies_clean.csv', index=False)
ratings.to_csv('data/ratings_clean.csv', index=False)
tags.to_csv('data/tags_clean.csv', index=False)

ratings = ratings.sample(frac=1.0)
train_lim = int(0.8 * len(ratings))
ratings[:train_lim].to_csv('data/ratings_train.csv', index=False)
ratings[train_lim:].to_csv('data/ratings_test.csv', index=False)
